+++ chaoshire/__init__.py
@@ -1,3 +1,3 @@
 """ChaosHire fairness-auditing package."""
 
-__version__ = "0.23.4"
+__version__ = "0.24.0"